agent/requirements.txt
faiss-cpu==1.9.0.post1
fastapi==0.115.6
langchain==0.3.20
langchain_community==0.3.19
langchain_core==0.3.43
langchain_openai==0.2.11
langchain_text_splitters==0.3.6
openpyxl==3.1.5
pandas==2.2.3
arize-phoenix==6.1.0
openinference-instrumentation-langchain==0.1.29
opensearch-py==2.4.2
psycopg2==2.9.10
pydantic==2.10.3
pypdf==5.1.0
python-dotenv==1.0.1
pymongo==4.10.1
PyYAML==6.0.2
unidecode==1.3.8
uvicorn==0.32.1
elasticsearch==8.17.2